Modified CartitemDTO and fixed Create Product BUG - "Cannot Invoke ColorDTO.getImages()"

+++ src/main/java/com/example/madrasdaapi/dto/commons/ProductLadderItem.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import lombok.ToString;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 @Data
 @ToString
@@ -11,8 +12,8 @@ public class ProductLadderItem implements Serializable {
      private Long id;
      private String name;
      private String imgUrl = "https://cdn.discordapp.com/attachments/1088691132210364437/1088693356714012692/image.png";
-     private Float profitAmount;
+     private BigDecimal profitAmount;
      private Long stocksSold;
      private Float profit;
-     private Float returnsContribution;
+     private Double returnsContribution;
 }
